Fix all problems

+++ 0x02-Session_authentication/api/v1/auth/session_db_auth.py
@@ -37,8 +37,9 @@ class SessionDBAuth(SessionExpAuth):
     def destroy_session(self, request=None):
         """Destroys the session associated with the request cookie."""
         session_id = self.session_cookie(request)
-        user_sessions = UserSession().search({"session_id": session_id})
-        if user_sessions:
-            user_sessions[0].delete()
-            return True
+        if session_id:
+            user_sessions = UserSession().search({"session_id": session_id})
+            if user_sessions:
+                user_sessions[0].delete()
+                return True
         return False
